new f32 api

+++ examples/custom_style.rs
@@ -1,6 +1,6 @@
 //PIPE me to a file!
 fn main() -> core::fmt::Result {
-    let x = (0..50).map(|x| (x as f64 / 50.0) * 10.0);
+    let x = (0..50).map(|x| (x as f32 / 50.0) * 10.0);
 
     let mut s = poloto::plot_with_html(
         "Demo: you can change the style of the svg file itself!",
